Reviewed-on: https://hexorix-qwuviqyyw.liara.run/mahdiimnzr/Admin_Panel-Logic_Flow/pulls/61

+++ src/core/utils/formatDate.js
@@ -3,7 +3,7 @@ const formatDate = (date) => {
   const event = new Date(date);
   const options = {
     day: "numeric",
-    month: "numeric",
+    month: "long",
     year: "numeric",
   };
   return event.toLocaleDateString("fa-IR", options);
